+++ blocklylib-core/src/main/assets/default/logic_blocks.json
@@ -21,7 +21,7 @@
     "nextStatement": null,
     "colour": "%{BKY_LOGIC_HUE}",
     "helpUrl": "%{BKY_CONTROLS_IF_HELPURL}",
-    "tooltip": "If a value is true, then do some statements."
+    "tooltip": "%{CONTROLS_IF_TOOLTIP_1}"
   },
   {
     "type": "controls_ifelse",
@@ -50,7 +50,7 @@
     "previousStatement": null,
     "nextStatement": null,
     "colour": "%{BKY_LOGIC_HUE}",
-    "tooltip": "If a value is true, then do the first block of statements. Otherwise, do the second block of statements.",
+    "tooltip": "%{BKY_CONTROLS_IF_TOOLTIP_2}",
     "helpUrl": "%{BKY_CONTROLS_IF_HELPURL}"
   },
   {
